pyproject.toml
[project]
name = "planweaver"
version = "0.1.0"
description = "Universal LLM Planning & Execution Engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "litellm>=1.35.0",
    "pyyaml>=6.0.1",
    "json-repair==0.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.0",
    "loguru>=0.7.3",
    "google-genai>=1.63.0",
    "pygithub>=2.8.1",
    "tavily-python>=0.7.21",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.22",
    "cachetools>=7.0.1",
    "slowapi>=0.1.9",
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.1",
]

[tool.pytest.ini_options]
minversion = "9.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-config",
    "--strict-markers",
    "--cov=src/planweaver",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
]
asyncio_mode = "strict"
markers = [
    "integration: exercises multiple services together",
    "e2e: exercises higher-level workflows",
]
filterwarnings = [
    "ignore:Support for class-based `config` is deprecated:DeprecationWarning:pydantic",
    "ignore:PyPDF2 is deprecated:DeprecationWarning:PyPDF2",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "planweaver.api.main",
    "planweaver.api.middleware",
    "planweaver.api.routers.context",
    "planweaver.cli",
    "planweaver.config",
    "planweaver.db.models",
    "planweaver.db.repositories",
    "planweaver.services.comparison_service",
    "planweaver.services.context_service",
    "planweaver.services.file_processor",
    "planweaver.services.github_analyzer",
    "planweaver.services.llm_gateway",
    "planweaver.services.optimizer_service",
    "planweaver.services.template_engine",
    "planweaver.services.web_search_service",
]
ignore_errors = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-ignore = ["E501"]
